% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculator.R
\name{ss_benefits}
\alias{ss_benefits}
\title{Computes social security benefits as a function of an individual's history
of earnings, retirement age, year in which they retired, and disability /
household status.}
\usage{
ss_benefits(ss_policy, earnings_history, retirement_age, retirement_year,
  current_years, disability_status = NULL, spousal_status = NULL)
}
\arguments{
\item{ss_policy}{A list containing parameters for social security policy. See
'Details'.}

\item{retirement_age}{The age at which an individual retires. Must be larger
than 62.}

\item{retirement_year}{The year in which an individual retires.}

\item{current_years}{A list of years in which to compute social security
benefits.}

\item{disability_status}{True if individual is on social security.}

\item{spousal_status}{True if individual is married.}
}
\value{
Vector of monthly social security benefit earnings, per year in
\code{current_year}.
}
\description{
Computes social security benefits as a function of an individual's history
of earnings, retirement age, year in which they retired, and disability /
household status.
}
\details{
The policy parameters consist of a wage index, so-called 'bend
points', retirement bonuses for retiring late, as well as cost-of-living
increases. The current values for these policies may be found on the social
security website, https://www.ssa.gov/oact/cola/Benefits.html. The
`ss_policy` list must contain the following items:
\itemize{
  \item{\code{wage_index}: }{Wage indices used to inflate past earnings.
  (See here: https://www.ssa.gov/oact/cola/AWIgrowth.html.)}
  \item{\code{bend_points}: }{So-called 'bend points'.
  (See here: https://www.ssa.gov/oact/cola/bendpoints.html.)}
  \item{\code{pia_percentages}: }{Discounts applied to benefits depending on age of
retirement.}
  \item{\code{cola_increases}: }{Annual cost of living increases applied to benefits.}
}
For formatting these inputs, see the data \code{ss_policy_defaults}, which
returns default values in the correct format, using current social security
administration policy as a guide. Load it with \code{data(ss_policy_defaults)}.
}

